Update DataGrid.rst
This is to fix a documentation bug listed in the bugs list as #39. 

In code examples, I changed css @import to <link> tags to fix the 404s on those CSS files. This seems to have been done at dojox/widget/Toaster.rst and the theming works fine there.

Please note I was not able to test it as I don't have a local rst test env. setup as yet.